ENG-1918 reuse config (#167)
ENG-1918:  If custom-config already exists ask if the user wants to keep it.  If so, don't delete it.  Requires user to re-use the same bucket for storage.  Uploads the pipeline directly to cloud storage, setting a pre-defined ID that can be reused on subsequent attempts.

NOTE:  If you've already installed with a previous installer, this will create a duplicate Deploy pipeline under the new ID.  I recommend you rename your old Deploy pipeline before running this as an upgrade, so you know which one to delete/not run later.